Cleanup run/stop scripts

+++ lw1/share/stop.sh
@@ -2,11 +2,14 @@
 
 # Usage: killDotNetApp <AppDllFilename>
 killDotNetApp() {
-    local PID_LIST=$(ps aux | grep $1 | head -n -1 | awk '{print $2}')
-    for pid in ${PID_LIST}; do {
-        echo "kill process ${pid} ($1)"
-        kill ${pid}
-    } done
+    local PROCESS_LIST=$(ps aux | grep $1)
+    if [ $(echo "${PROCESS_LIST}" | wc -l) > 1 ]; then
+        local PID_LIST=$(echo ${PROCESS_LIST} | head -n -1 | awk '{print $2}')
+        for pid in ${PID_LIST}; do {
+            echo "kill process ${pid} ($1)"
+            kill ${pid}
+        } done
+    fi
 }
 
 killDotNetApp "Frontend.dll"
